+++ packages/vexide-devices/src/smart/rotation.rs
@@ -118,7 +118,7 @@ impl RotationSensor {
         // behavior on our end without ever touching the status code.
         //
         // For more information: <https://www.vexforum.com/t/rotation-sensor-bug-workaround-on-vexos-1-1-0/96577/2>
-        if new_direction != self.direction()? {
+        if new_direction != self.direction() {
             self.direction_offset = self.position()?;
             self.raw_direction_offset = Position::from_ticks(
                 i64::from(unsafe { vexDeviceAbsEncPositionGet(self.device) }),
@@ -152,14 +152,8 @@ impl RotationSensor {
     }
 
     /// Gets whether the rotation sensor is reversed.
-    ///
-    /// # Errors
-    ///
-    /// An error is returned if an optical sensor is not currently connected to the smart port.
-    pub fn direction(&self) -> Result<Direction, PortError> {
-        self.validate_port()?;
-
-        Ok(self.direction)
+    pub const fn direction(&self) -> Direction {
+        self.direction
     }
 
     /// Get the total number of degrees rotated by the sensor based on direction.