+++ packages/vexide-macro/src/lib.rs
@@ -127,7 +127,7 @@ fn make_entrypoint(inner: &ItemFn, opts: MacroOpts) -> proc_macro2::TokenStream 
 /// # use core::fmt::Write;
 /// #[vexide::main]
 /// async fn main(mut peripherals: Peripherals) {
-///     write!(peripherals.screen, "Hello, vexide!").unwrap();
+///     write!(peripherals.display, "Hello, vexide!").unwrap();
 /// }
 /// ```
 ///